Don't store per-request info as server attributes
The .setup() method was too complicated, and relied on a misfeature of
Rakudo which confuses attributes and methods (using .can() to test for
attribute existence).  See:
    http://irclog.perlgeek.de/perl6/2011-07-11#i_4094525

Now, .setup() is simply a stub that's meant to be overridden, and the
subclass can handle that data appropriately.

Move the call to .headers() after the call to .setup(), to make
it easier to set up a request environment before processing the headers.

Also, fixed examples/02-simple-small.pl6, which had two methods which
were never called (replace them with a .setup() method instead).